+++ packages/gasket-plugin-https/test/index.test.js
@@ -33,6 +33,7 @@ describe('Plugin', () => {
   it('has expected hooks', () => {
     const expected = [
       'actions',
+      'create',
       'metadata'
     ];