+++ packages/gasket-plugin-nextjs/test/create.test.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const { devDependencies } = require('../package');
+const { name, version, devDependencies } = require('../package');
 
 describe('create hook', () => {
   let mockContext;
@@ -14,7 +14,8 @@ describe('create hook', () => {
       },
       files: { add: jest.fn() },
       gasketConfig: {
-        add: jest.fn()
+        add: jest.fn(),
+        addPlugin: jest.fn()
       }
     };
   });
@@ -54,20 +55,35 @@ describe('create hook', () => {
     );
   });
 
+  it('adds itself to the dependencies', async function () {
+    await plugin.hooks.create.handler({}, mockContext);
+
+    expect(mockContext.pkg.add).toHaveBeenCalledWith('dependencies',
+      expect.objectContaining({ [name]: `^${version}` })
+    );
+  });
+
   it('adds appropriate dependencies', async function () {
     await plugin.hooks.create.handler({}, mockContext);
 
-    expect(mockContext.pkg.add).toHaveBeenCalledWith('dependencies', {
-      '@gasket/assets': devDependencies['@gasket/assets'],
-      '@gasket/nextjs': devDependencies['@gasket/nextjs'],
-      'next': devDependencies.next,
-      'prop-types': devDependencies['prop-types'],
-      'react': devDependencies.react,
-      'react-dom': devDependencies['react-dom']
-    });
+    expect(mockContext.pkg.add).toHaveBeenCalledWith('dependencies',
+      expect.objectContaining({
+        '@gasket/assets': devDependencies['@gasket/assets'],
+        '@gasket/nextjs': devDependencies['@gasket/nextjs'],
+        'next': devDependencies.next,
+        'prop-types': devDependencies['prop-types'],
+        'react': devDependencies.react,
+        'react-dom': devDependencies['react-dom']
+      }));
+  });
+
+  it('add plugin import to the gasket file', async function () {
+    await plugin.hooks.create.handler({}, mockContext);
+    expect(mockContext.gasketConfig.addPlugin).toHaveBeenCalledWith('pluginNextjs', name);
   });
 
   it('adds appropriate devDependencies', async function () {
+    mockContext.nextDevProxy = true;
     await plugin.hooks.create.handler({}, mockContext);
     expect(mockContext.pkg.add).toHaveBeenCalledWith('devDependencies', {
       nodemon: devDependencies.nodemon
@@ -118,8 +134,8 @@ describe('create hook', () => {
   });
 
   it('adds next config & server files', async function () {
+    mockContext.nextDevProxy = true;
     await plugin.hooks.create.handler({}, mockContext);
-
     expect(mockContext.files.add).toHaveBeenCalledWith(
       `${root}/../generator/next/*`
     );
@@ -133,7 +149,7 @@ describe('create hook', () => {
       'build': 'next build',
       'start': 'next start',
       'start:local': 'next start & GASKET_ENV=local node server.js',
-      'local': 'next dev & nodemon server.js'
+      'local': 'next dev'
     });
   });