+++ packages/gasket-plugin-winston/lib/index.js
@@ -3,14 +3,16 @@
 /// <reference types="@gasket/plugin-metadata" />
 
 const { createLogger, format, transports } = require('winston');
-const { name, dependencies } = require('../package.json');
+const { name, version, dependencies } = require('../package.json');
 
 /** @type {import('@gasket/engine').Plugin} */
 const plugin = {
   name,
   hooks: {
-    async create(gasket, { pkg }) {
+    async create(gasket, { pkg, gasketConfig }) {
+      gasketConfig.addPlugin('pluginWinston', '@gasket/plugin-winston');
       pkg.add('dependencies', {
+        [name]: `^${version}`,
         winston: dependencies.winston
       });
     },