+++ packages/gasket-preset-nextjs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # `@gasket/preset-nextjs`
 
+
+
+- Convert to ESM ([#736])
+- Add preset hooks `presetPrompt` and `presetConfig`
+- Add tests
+- Update dependencies
+
 ### 6.10.0
 
 - TypeScript definitions for config & lifecycles ([#311])
@@ -17,3 +24,4 @@
 - Initial release.
 
 [#311]: https://github.com/godaddy/gasket/pull/311
+[#736]: https://github.com/godaddy/gasket/pull/736